Move Reporter into its own file
